Don't add a header to empty files
Python requries package folders to have an '__init__.py' file, but usually this file is empty.
Since there is no code in the file adding a copyright header seemed to me to be a bit non-sensicle.
I suppose this could be a flag, but I really can't think of why anyone should need a copyright header for an empty file?